restructured arquin module, makes Device and Module FrozenClass
